Better doc and packaging
