+++ tools/proto-dart-plugin/src/main/java/io/spine/code/dart/lexeme/Import.java
@@ -20,7 +20,7 @@
 
 package io.spine.code.dart.lexeme;
 
-import io.spine.code.dart.FileName;
+import io.spine.code.fs.dart.FileName;
 
 import static com.google.common.base.Preconditions.checkNotNull;
