+++ tools/tool-base/src/main/java/io/spine/code/fs/js/FileName.java
@@ -34,8 +34,8 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import static io.spine.util.Preconditions2.checkNotEmptyOrBlank;
 
 /**
- * A name of a JavaScript file, which is generated under the
- * the root directory with compiled Protobuf files.
+ * A name of a JavaScript file, which is generated under the the root directory with compiled
+ * Protobuf files.
  */
 public final class FileName extends AbstractFileName<FileName> {
